# packages/contracts/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "JB API",
  "version": "0.1.0"
 },
 "paths": {
  "/api/v1/companies": {
   "post": {
    "tags": [
     "companies"
    ],
    "summary": "Create Company",
    "description": "Create a company from observables, run the valuation ensemble, open its market.\nRole B's discovery pipeline calls this (or engine.create_company directly) per company.",
    "operationId": "create_company_api_v1_companies_post",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CompanyIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Company"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "companies"
    ],
    "summary": "List Companies",
    "operationId": "list_companies_api_v1_companies_get",
    "parameters": [
     {
      "name": "state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "State"
      }
     },
     {
      "name": "category",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Category"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "pattern": "^(v0|confidence|last|name)$",
       "default": "v0",
       "title": "Sort"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/CompanyCard"
         },
         "title": "Response List Companies Api V1 Companies Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/companies/{cid}": {
   "get": {
    "tags": [
     "companies"
    ],
    "summary": "Get Company",
    "operationId": "get_company_api_v1_companies__cid__get",
    "parameters": [
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Cid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Company"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/companies/valuation/preview": {
   "post": {
    "tags": [
     "companies"
    ],
    "summary": "Preview Valuation",
    "description": "Run the ensemble without creating anything. Handy for the UI and for calibration.",
    "operationId": "preview_valuation_api_v1_companies_valuation_preview_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CompanyIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Valuation"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/companies/{cid}/appraise": {
   "post": {
    "tags": [
     "companies"
    ],
    "summary": "Appraise",
    "description": "Grok researches the business on the web and gives a value with sources; K2 gives a second number.\nBoth feed the ensemble's `llm` estimator and the valuation is recomputed. The market prior is\nre-anchored only if nothing has traded yet; after that the belief already reflects the market.",
    "operationId": "appraise_api_v1_companies__cid__appraise_post",
    "parameters": [
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Cid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Company"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/companies/{cid}/ask": {
   "post": {
    "tags": [
     "companies"
    ],
    "summary": "Ask Owner",
    "description": "Talk to the AI version of the owner, grounded in the file (profile, valuation, sources).",
    "operationId": "ask_owner_api_v1_companies__cid__ask_post",
    "parameters": [
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Cid"
      }
     },
     {
      "name": "x-demo-user",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Demo-User"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AskIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/acquire/{cid}/start": {
   "post": {
    "tags": [
     "acquire"
    ],
    "summary": "Start",
    "operationId": "start_api_v1_acquire__cid__start_post",
    "parameters": [
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Cid"
      }
     },
     {
      "name": "x-demo-user",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Demo-User"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Acquisition"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/acquire/{acq_id}": {
   "get": {
    "tags": [
     "acquire"
    ],
    "summary": "Get",
    "operationId": "get_api_v1_acquire__acq_id__get",
    "parameters": [
     {
      "name": "acq_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Acq Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Acquisition"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "Acquisition": {
    "properties": {
     "acquisition_id": {
      "type": "string",
      "title": "Acquisition Id"
     },
     "market_id": {
      "type": "string",
      "title": "Market Id"
     },
     "loi_md": {
      "type": "string",
      "title": "Loi Md"
     },
     "checklist": {
      "items": {
       "$ref": "#/components/schemas/ChecklistItem"
      },
      "type": "array",
      "title": "Checklist"
     },
     "status": {
      "type": "string",
      "enum": [
       "draft",
       "sent",
       "closed"
      ],
      "title": "Status",
      "default": "draft"
     }
    },
    "type": "object",
    "required": [
     "acquisition_id",
     "market_id",
     "loi_md",
     "checklist"
    ],
    "title": "Acquisition"
   },
...
   "AskIn": {
    "properties": {
     "question": {
      "type": "string",
      "title": "Question"
     },
     "history": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "History",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "question"
    ],
    "title": "AskIn"
   },
...
   "Company": {
    "properties": {
     "_id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "category": {
      "type": "string",
      "title": "Category"
     },
     "naics_guess": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Naics Guess"
     },
     "address": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "state": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "State"
     },
     "lat": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lat"
     },
     "lng": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lng"
     },
     "website": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Website"
     },
     "phone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     },
     "rating": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rating"
     },
     "review_count": {
      "type": "integer",
      "title": "Review Count"
     },
     "founded_year": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Founded Year"
     },
     "owners": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Owners"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "financials": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Financials"
       },
       {
        "type": "null"
       }
      ]
     },
     "valuation": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Valuation"
       },
       {
        "type": "null"
       }
      ]
     },
     "sources": {
      "items": {
       "$ref": "#/components/schemas/Source"
      },
      "type": "array",
      "title": "Sources"
     },
     "status": {
      "type": "string",
      "enum": [
       "stub",
       "ready",
       "failed"
      ],
      "title": "Status"
     },
     "created_at": {
      "type": "string",
      "title": "Created At"
     },
     "market": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/MarketSummary"
       },
       {
        "type": "null"
       }
      ]
     },
     "observables": {
      "additionalProperties": true,
      "type": "object",
      "title": "Observables"
     }
    },
    "type": "object",
    "required": [
     "_id",
     "name",
     "category",
     "naics_guess",
     "address",
     "city",
     "state",
     "lat",
     "lng",
     "website",
     "phone",
     "rating",
     "review_count",
     "founded_year",
     "owners",
     "description",
     "financials",
     "valuation",
     "sources",
     "status",
     "created_at",
     "market",
     "observables"
    ],
    "title": "Company"
   },
   "CompanyCard": {
    "properties": {
     "_id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "category": {
      "type": "string",
      "title": "Category"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "state": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "State"
     },
     "rating": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rating"
     },
     "review_count": {
      "type": "integer",
      "title": "Review Count"
     },
     "bid": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bid"
     },
     "ask": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ask"
     },
     "last": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last"
     },
     "indicative_price": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Indicative Price"
     },
     "v0_per_share": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "V0 Per Share"
     },
     "confidence": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Confidence"
     },
     "status": {
      "type": "string",
      "enum": [
       "stub",
       "ready",
       "failed"
      ],
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "_id",
     "name",
     "category",
     "city",
     "state",
     "rating",
     "review_count",
     "bid",
     "ask",
     "last",
     "indicative_price",
     "v0_per_share",
     "confidence",
     "status"
    ],
    "title": "CompanyCard"
   },
   "CompanyIn": {
    "properties": {
     "category": {
      "type": "string",
      "title": "Category",
      "default": "default"
     },
     "state": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "State"
     },
     "revenue": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Revenue"
     },
     "sde": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sde"
     },
     "asking_price": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Asking Price"
     },
     "employees": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Employees"
     },
     "rating": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rating"
     },
     "review_count": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Review Count"
     },
     "years_operating": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Years Operating"
     },
     "owner_operated": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Owner Operated"
     },
     "llm_estimate": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Llm Estimate"
     },
     "llm_confidence": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Llm Confidence"
     },
     "llm2_estimate": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Llm2 Estimate"
     },
     "machines": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Machines"
     },
     "sources": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Sources"
     },
     "id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "address": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "website": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Website"
     },
     "phone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     },
     "naics_guess": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Naics Guess"
     },
     "lat": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lat"
     },
     "lng": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lng"
     },
     "owners": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Owners"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "CompanyIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Source": {
    "properties": {
     "url": {
      "type": "string",
      "title": "Url"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "snippet": {
      "type": "string",
      "title": "Snippet"
     },
     "fetched_at": {
      "type": "string",
      "title": "Fetched At"
     }
    },
    "type": "object",
    "required": [
     "url",
     "title",
     "snippet",
     "fetched_at"
    ],
    "title": "Source"
   },
...
   "Valuation": {
    "properties": {
     "v0": {
      "type": "number",
      "title": "V0"
     },
     "sigma": {
      "type": "number",
      "title": "Sigma"
     },
     "low": {
      "type": "number",
      "title": "Low"
     },
     "high": {
      "type": "number",
      "title": "High"
     },
     "method": {
      "type": "string",
      "title": "Method"
     },
     "disagreement": {
      "type": "number",
      "title": "Disagreement"
     },
     "estimates": {
      "items": {
       "$ref": "#/components/schemas/Estimate"
      },
      "type": "array",
      "title": "Estimates"
     },
     "as_of": {
      "type": "string",
      "title": "As Of"
     }
    },
    "type": "object",
    "required": [
     "v0",
     "sigma",
     "low",
     "high",
     "method",
     "disagreement",
     "estimates",
     "as_of"
    ],
    "title": "Valuation"
   }
  }
 }
}